requirements.txt
pyTelegramBotAPI
youtube_dl
pytube
youtube-search-python